refactor: move process_stage to the enums folder
