+++ lib/core/domain/enums/process_stage.dart
